test: add GetMoviePopularWithProviderTMDBTestApiUnitTest
